Initial filesystem based Vault setup
This shows a simple startup script and verification step for a single
node Vault server that is backed by the filesystem.

Clearly there are many ways this could be improved, like by building
most of this into the base image, using Ansible for the setup,
provisioning certificates and using TLS, and, of course, setting up a
backend that is actually highly available.

However, for now this is a good example of how to use Cloudless to
install a real service, and this might see some use as a "bootstrap"
Vault (because when you're setting up a real Consul/Vault cluster, how
do you get your initial set of secrets?).